# Fix bug in performing sweeps