This merge fixes a bug when performing sweeps. Previously only half of the sweeps were executed. After this merge, all sweeps are executed giving more accurate performance data.