@tomdeakin 2018-01-08
Thanks for the pull request. I've had a look at the code and have a couple of comments.

The MPI communications are already timed as part of the comms module and printed out by the driver code. Therefore the new timers for this do not need to be added as per your commit ae6fa8b16829066c57e14fc252

It looks like you overwrite the sweep times array for each `ntimes` iteration, and therefore only the final sweep times are written out at the end. Is this your intended behaviour? Also, I would be keen to know if including timers for each sweep slowed the code down at all - did you notice anything?

Good spot with the iteration bug in your commit 4723304d1de869cada594! The 3D version of `mega-sweep` is definitely less mature than the 2D version. I'm happy to accept this change but it is better done as a separate pull request to the timers.