@douglasjacobsen 2018-01-08
@tomdeakin I'll change this PR to be the timers, and make a separate one about the bug fix, since the majority of your comment relates to the timers. I had only included them together because adding the per-sweep timers allows one to easily see that before the bug fix, only half of the sweeps are executed.

The MPI communications are only timed here to remove from the sweep timers. The time more instructions than are included in the comms timers, but they are similar. I can swap to the comms timers if you would prefer that. In the end, it'll just make the per-sweep timers a bit larger than they should be, but they will still illustrate the issue that I want to point out.

Upon inspection, it looks like the sweep timers should be an accumulation of per-sweep time over all ntimes. `sweep_times` is only initialized to zero outside of the `ntimes` loop (shortly after allocation) and within `kernel3d.f90` it is written to accumulate the sweep time into the array rather than to overwrite it. I might be missing something though, so if you see somewhere that there is an issue, please let me know.

As far as the performance impact of adding the timers goes, here is some performance data from before and after the timers. NOTE: I've included the fix for the different sweep directions in both versions, so all sweeps are performing their work now.

## Run command
```
└─▪ OMP_NUM_THREADS=1 ./mega-sweep3d --nang 48 --ng 16 --nx 720 --ny 2 --nz 2 --chunk 24 --ntimes 100
```

## Before
```
Summary
 Fastest iteration (s):    0.162029982
 Slowest iteration (s)     0.172363997

 Time in MPI_Recv (s):        0.118762493  0.7%
 Time in MPI_Wait (s):        0.127095938  0.8%
 Compute time (s):           16.104736090 98.5%
 Remaining time (s):          0.000173330  0.0%

 Runtime (s):                16.350767851
```

## After
```
Summary
 Fastest iteration (s):    0.161821127
 Slowest iteration (s)     0.170040846

 Time in MPI_Recv (s):        0.104841232  0.6%
 Time in MPI_Wait (s):        0.113211632  0.7%
 Compute time (s):           16.102960587 98.7%
 Remaining time (s):          0.000149488  0.0%

 Runtime (s):                16.321162939
```

Based on this test, there is little to no negative performance impact of including the per sweep timers in the run.

Let me know how you'd like to proceed with the timer PR, and I'll submit the bug fix PR asap.