@douglasjacobsen 2018-01-08
Sorry, this PR will become the bugfix and I'll make a new one for timers. But I'll refer to these commits in that one. Just since the branch name is bugfix for this PR.